+++ src/WeightRedistributor.php
@@ -117,8 +117,9 @@ class WeightRedistributor implements LoggerAwareInterface
                 continue; // moving this item would harm more than help
             }
 
-            $newLighterBoxes = $this->doVolumeRepack(array_merge($underWeightBoxItems, [$overWeightItem]), $underWeightBox->box);
-            if ($newLighterBoxes->count() !== 1) {
+            $packableItems = array_merge($underWeightBoxItems, [$overWeightItem]);
+            $newLighterBoxes = $this->doVolumeRepack($packableItems, $underWeightBox->box);
+            if ($this->isEverythingPacked($newLighterBoxes, $packableItems) === false) {
                 continue; // only want to move this item if it still fits in a single box
             }
 
@@ -135,7 +136,7 @@ class WeightRedistributor implements LoggerAwareInterface
 
             unset($overWeightBoxItems[$key]);
             $newHeavierBoxes = $this->doVolumeRepack($overWeightBoxItems, $overWeightBox->box);
-            if (count($newHeavierBoxes) !== 1) {
+            if ($this->isEverythingPacked($newHeavierBoxes, $overWeightBoxItems) === false) {
                 assert(true, 'Could not pack n-1 items into box, even though n were previously in it');
                 continue;
             }
@@ -151,6 +152,20 @@ class WeightRedistributor implements LoggerAwareInterface
         }
 
         return $anyIterationSuccessful;
+    }
+
+    private function isEverythingPacked(PackedBoxList $result, array $expectedItems): bool
+    {
+        if ($result->count() !== 1) {
+            return false;
+        }
+
+        $packedItems = [];
+        foreach ($result->top()->items as $packedItem) {
+            $packedItems[] = $packedItem->item;
+        }
+
+        return count($packedItems) === count($expectedItems);
     }
 
     /**